migrate the rest of the auth endpoints to Axum (#575)
