+++ dim-core/src/routes/mod.rs
@@ -21,7 +21,6 @@ pub mod auth;
 pub mod dashboard;
 pub mod general;
 pub mod host;
-pub mod invites;
 pub mod library;
 pub mod media;
 pub mod mediafile;
