+++ dim-web/src/routes/auth.rs
@@ -56,6 +56,187 @@ impl IntoResponse for AuthError {
     }
 }
 
+/// # GET `/api/v1/auth/invites`
+/// Method will retrieve and return all invite tokens in the database.
+///
+/// # Authorization
+/// This route requires a valid authentication token to be supplied. The token must have `owner`
+/// permissions.
+///
+/// # Request
+/// ## Example
+/// ```text
+/// curl -X GET http://127.0.0.1:8000/api/v1/auth/invites -H "Authorization: ...."
+/// ```
+///
+/// # Response
+/// The route will return a response with the following schema
+/// ```no_compile
+/// [
+///   {
+///     "id": String,
+///     "created": i64,
+///     "claimed_by": Option<String>,
+///   },
+///   ...
+/// ]
+/// ```
+///
+/// ## Example
+/// ```no_compile
+/// [
+///   {
+///     "id": "079a38b4-d39f-4a9e-9a18-964f225b75d3",
+///     "created": 1638708402,
+///     "claimed_by": "admin"
+///   },
+///   {
+///     "id": "844caa7b-f54f-a9ea-4444-555555555555",
+///     "created": 1640000000,
+///   }
+/// ]
+/// ```
+///
+/// # Errors
+/// * [`Unauthorized`] - Returned if the authentication token lacks `owner` permissions
+///
+/// [`Unauthorized`]: crate::errors::DimError::Unauthorized
+pub async fn get_all_invites(
+    Extension(user): Extension<User>,
+    extract::State(conn): extract::State<DbConnection>,
+) -> Result<axum::response::Response, AuthError> {
+    let mut tx = conn.read().begin().await.map_err(DatabaseError::from)?;
+    if user.has_role("owner") {
+        #[derive(serde::Serialize)]
+        struct Row {
+            id: String,
+            created: i64,
+            claimed_by: Option<String>,
+        }
+
+        // FIXME: LEFT JOINs cause sqlx::query! to panic, thus we must get tokens in two queries.
+        // TODO: Move these into database.
+        // TODO: We silently drop db errors here, we should probably change this.
+        let mut row = sqlx::query_as!(
+            Row,
+            r#"SELECT invites.id, invites.date_added as created, NULL as "claimed_by: _"
+                FROM invites
+                WHERE invites.id NOT IN (SELECT users.claimed_invite FROM users)
+                ORDER BY created ASC"#
+        )
+        .fetch_all(&mut tx)
+        .await
+        .unwrap_or_default();
+
+        row.append(
+            &mut sqlx::query_as!(
+                Row,
+                r#"SELECT invites.id, invites.date_added as created, users.username as "claimed_by: Option<String>"
+            FROM  invites
+            INNER JOIN users ON users.claimed_invite = invites.id"#
+            )
+            .fetch_all(&mut tx)
+            .await
+            .unwrap_or_default(),
+        );
+
+        return Ok(axum::response::Json(json!(&row)).into_response());
+    }
+
+    Err(AuthError::InvalidCredentials)
+}
+
+
+/// # POST `/api/v1/auth/new_invite`
+/// Method will generate and return a new invite token.
+///
+/// # Authorization
+/// This route requires a valid authentication token to be supplied. The token must have `owner`
+/// permissions.
+///
+/// # Request
+/// ## Example
+/// ```text
+/// curl -X POST http://127.0.0.1:8000/api/v1/auth/new_invite -H "Authorization: ...."
+/// ```
+///
+/// # Response
+/// The route will return a response with the following schema
+/// ```no_compile
+/// {
+///   "token": String,
+/// }
+/// ```
+///
+/// ## Example
+/// ```no_compile
+/// {
+///   "token": "844caa7b-f54f-a9ea-4444-555555555555",
+/// }
+/// ```
+///
+/// # Errors
+/// * [`Unauthorized`] - Returned if the authentication token lacks `owner` permissions
+///
+/// [`Unauthorized`]: crate::errors::DimError::Unauthorized
+pub async fn generate_invite(
+    Extension(user): Extension<User>,
+    extract::State(conn): extract::State<DbConnection>,
+) -> Result<axum::response::Response, AuthError> {
+    if !user.has_role("owner") {
+        return Err(AuthError::InvalidCredentials);
+    }
+
+    let mut lock = conn.writer().lock_owned().await;
+    let mut tx = dim_database::write_tx(&mut lock).await.map_err(DatabaseError::from)?;
+
+    let token = Login::new_invite(&mut tx).await?;
+
+    tx.commit().await.map_err(DatabaseError::from)?;
+
+    Ok(axum::response::Json(json!({ "token": token })).into_response())
+}
+
+
+/// # DELETE `/api/v1/auth/token/:token`
+/// Method will revoke the supplied token.
+///
+/// # Authorization
+/// This route requires a valid authentication token to be supplied. The token must have `owner`
+/// permissions.
+///
+/// # Request
+/// This request takes in a route parameter which is the token we want to delete.
+/// ## Example
+/// ```text
+/// curl -X DELETE http://127.0.0.1:8000/api/v1/auth/token/844caa7b-f54f-a9ea-4444-555555555555 -H "Authorization: ...."
+/// ```
+///
+/// # Response
+/// If the token was successfully deleted, this route will return `200 0K`.
+///
+/// # Errors
+/// * [`Unauthorized`] - Returned if the authentication token lacks `owner` permissions
+///
+/// [`Unauthorized`]: crate::errors::DimError::Unauthorized
+pub async fn delete_token(
+    Extension(user): Extension<User>,
+    extract::State(conn): extract::State<DbConnection>,
+    extract::Path(token): extract::Path<String>,
+) -> Result<impl IntoResponse, AuthError> {
+    if !user.has_role("owner") {
+        return Err(AuthError::InvalidCredentials);
+    }
+
+    let mut lock = conn.writer().lock_owned().await;
+    let mut tx = dim_database::write_tx(&mut lock).await.map_err(DatabaseError::from)?;
+    Login::delete_token(&mut tx, token).await?;
+    tx.commit().await.map_err(DatabaseError::from)?;
+
+    Ok(StatusCode::OK)
+}
+
+
 /// # GET `/api/v1/user`
 /// Method returns metadata about the currently logged in user.
 ///
